+++ sql/core/src/test/scala/org/apache/spark/sql/sources/InsertSuite.scala
@@ -1568,15 +1568,18 @@ class InsertSuite extends DataSourceTest with SharedSparkSession {
     }
 
     // This represents one test configuration over a data source.
-    case class Config(dataSource: String, sqlConf: Seq[Option[(String, String)]] = Seq())
+    case class Config(
+      dataSource: String,
+      sqlConf: Seq[Option[(String, String)]] = Seq())
+    // Run the test several times using each configuration.
     Seq(
       Config(dataSource = "json",
         Seq(
           Some(SQLConf.JSON_GENERATOR_IGNORE_NULL_FIELDS.key -> "false"))),
       Config(dataSource = "csv",
         Seq(
-          Some(SQLConf.CSV_PARSER_COLUMN_PRUNING.key -> "false"),
-          None))
+          None,
+          Some(SQLConf.CSV_PARSER_COLUMN_PRUNING.key -> "false")))
     ).foreach { config: Config =>
       config.sqlConf.foreach {
         _.map { kv: (String, String) =>
